mcp_auth/requirements.txt
fastmcp
PyJWT
